refactor(server): Split server.rs into per-concern modules and add LLMClient trait
Break the 3000-line `src/server.rs` into a `server/` directory mirroring
the Go ADK's per-file layout (`agent.rs`, `agent_builder.rs`,
`task_handler.rs`, `server_core.rs`, `server_builder.rs`, `protocol.rs`,
`agent_card.rs`, `agent_toolbox.rs`, `errors.rs`). Public API surface
is unchanged.

Introduce `LLMClient` trait + `OpenAICompatibleLLMClient` struct
(`agent_llm_client.rs`) matching the Go ADK's interface. `Agent` now
holds `Arc<dyn LLMClient>`; `run_tool_loop` and `stream_agent_deltas`
dispatch through the trait. `AgentBuilder` gains `with_llm_client` for
custom backends; the default `OpenAICompatibleLLMClient` is built from
`AgentConfig` (including retry behaviour). Default task handlers no
longer require a separate `gateway_url` — the agent owns the URL via
its LLM client.
