+++ examples/static-agent-card/server/main.rs
@@ -6,10 +6,14 @@ use tracing::{error, info};
 async fn main() -> Result<(), Box<dyn std::error::Error>> {
     tracing_subscriber::fmt().init();
 
-    let config = Config::from_env()?;
+    let mut config = Config::from_env()?;
 
     let gateway_url = env::var("INFERENCE_GATEWAY_URL")
         .unwrap_or_else(|_| "http://localhost:8080/v1".to_string());
+
+    if config.agent_config.base_url.is_none() {
+        config.agent_config.base_url = Some(gateway_url.clone());
+    }
 
     info!("Starting A2A server with Inference Gateway SDK...");
     info!("Gateway URL: {}", gateway_url);
